Switched over from os.system to subprocess.Popen for running mplayer.

+++ reddit-radio.py
@@ -3,8 +3,6 @@ import subprocess
 import os
 import sys
 
-#test
-
 #set the subreddit name from the passed argument or set the default
 if 1 < len(sys.argv):
   sub = str(sys.argv[1])
@@ -19,10 +17,13 @@ domain_white_list = ['youtube.com', 'soundcloud.com', 'mixcloud.com']
 #limit of entries retrieved from reddit
 data_limit = 50
 
-#base data for reddit api
-after = 'null'
-before = 'null'
-count = 0
+# Reddit Parameters.
+params = {
+	'after': 0,
+	'before': 0,
+	'count': 0,
+	'limit': '100'
+}
 
 # delete temporary file if it exists
 #if os.path.isfile('temp.radio'):
@@ -32,19 +33,23 @@ session = requests.Session()
 session.headers.update({"User-Agent": "Reddit Radio/0.1",})
 
 while 1:
-	payload = {'after': after, 'before': before, 'count': count, 'limit': '100'}
-	r = session.post(base_subreddit, data=payload)
+	response = session.post(base_subreddit, data=params)
 
-	after = r.json()['data']['after']
-	before = r.json()['data']['before']
-	count += len(r.json()['data']['children']);
+	params['after'] = response.json()['data']['after']
+	params['before'] = response.json()['data']['before']
+	params['count'] += len(response.json()['data']['children'])
 
-	for i in r.json()['data']['children']:
+	for i in response.json()['data']['children']:
 		if i['data']['domain'] in domain_white_list:
-			os.system('python youtube-dl -q -f bestaudio -o temp.radio ' + i['data']['url'])
-			print 'Playing song... ' + i['data']['title']
-			os.system('mpv -really-quiet temp.radio')
-			os.remove('temp.radio')
+			os.system("youtube-dl -q -f bestaudio -o temp.radio " + i['data']['url'])
+			print '[PLAYING] ' + i['data']['title']
+			process = subprocess.Popen("mplayer -really-quiet temp.radio", stdin=subprocess.PIPE, shell=True)
+			while True:
+				key = raw_input()
+				if key == 'q':
+					process.kill()
+					os.remove('temp.radio')
+					break
 
 	if count >= data_limit:
 		break
